third_party/blink/web_tests/platform/generic/http/tests/security/cross-frame-access-get-expected.txt

This test checks cross-frame access security (rdar://problem/5251309).


----- tests for getting of allowed properties -----


----- tests for getting of allowed Functions -----

PASS: canGet('targetWindow.blur') should be 'true' and is.
PASS: canGet('targetWindow.close') should be 'true' and is.
PASS: canGet('targetWindow.focus') should be 'true' and is.

----- tests for getting of allowed Attributes -----

PASS: canGet('targetWindow.closed') should be 'true' and is.
PASS: canGet('targetWindow.frames') should be 'true' and is.
PASS: canGet('targetWindow.length') should be 'true' and is.
PASS: canGet('targetWindow.opener') should be 'true' and is.
PASS: canGet('targetWindow.parent') should be 'true' and is.
PASS: canGet('targetWindow.self') should be 'true' and is.
PASS: canGet('targetWindow.top') should be 'true' and is.
PASS: canGet('targetWindow.window') should be 'true' and is.

----- tests for getting of not allowed properties -----


----- tests for getting of not allowed Constructors -----

PASS: canGet('targetWindow.Attr') should be 'false' and is.
PASS: canGet('targetWindow.Audio') should be 'false' and is.
PASS: canGet('targetWindow.CDATASection') should be 'false' and is.
PASS: canGet('targetWindow.CSSRule') should be 'false' and is.
PASS: canGet('targetWindow.CSSStyleDeclaration') should be 'false' and is.
PASS: canGet('targetWindow.CharacterData') should be 'false' and is.
PASS: canGet('targetWindow.Comment') should be 'false' and is.
PASS: canGet('targetWindow.DOMException') should be 'false' and is.
PASS: canGet('targetWindow.DOMImplementation') should be 'false' and is.
PASS: canGet('targetWindow.DOMParser') should be 'false' and is.
PASS: canGet('targetWindow.Document') should be 'false' and is.
PASS: canGet('targetWindow.DocumentFragment') should be 'false' and is.
PASS: canGet('targetWindow.DocumentType') should be 'false' and is.
PASS: canGet('targetWindow.Element') should be 'false' and is.
PASS: canGet('targetWindow.EntityReference') should be 'false' and is.
PASS: canGet('targetWindow.EvalError') should be 'false' and is.
PASS: canGet('targetWindow.Event') should be 'false' and is.
PASS: canGet('targetWindow.HTMLAnchorElement') should be 'false' and is.
PASS: canGet('targetWindow.HTMLAudioElement') should be 'false' and is.
PASS: canGet('targetWindow.HTMLAreaElement') should be 'false' and is.
PASS: canGet('targetWindow.HTMLBRElement') should be 'false' and is.
PASS: canGet('targetWindow.HTMLBaseElement') should be 'false' and is.
PASS: canGet('targetWindow.HTMLBodyElement') should be 'false' and is.
PASS: canGet('targetWindow.HTMLButtonElement') should be 'false' and is.
PASS: canGet('targetWindow.HTMLCanvasElement') should be 'false' and is.
PASS: canGet('targetWindow.HTMLDListElement') should be 'false' and is.
PASS: canGet('targetWindow.HTMLDirectoryElement') should be 'false' and is.
PASS: canGet('targetWindow.HTMLDivElement') should be 'false' and is.
PASS: canGet('targetWindow.HTMLDocument') should be 'false' and is.
PASS: canGet('targetWindow.HTMLElement') should be 'false' and is.
PASS: canGet('targetWindow.HTMLFieldSetElement') should be 'false' and is.
PASS: canGet('targetWindow.HTMLFontElement') should be 'false' and is.
PASS: canGet('targetWindow.HTMLFormElement') should be 'false' and is.
PASS: canGet('targetWindow.HTMLFrameElement') should be 'false' and is.
PASS: canGet('targetWindow.HTMLFrameSetElement') should be 'false' and is.
PASS: canGet('targetWindow.HTMLHRElement') should be 'false' and is.
PASS: canGet('targetWindow.HTMLHeadElement') should be 'false' and is.
PASS: canGet('targetWindow.HTMLHeadingElement') should be 'false' and is.
PASS: canGet('targetWindow.HTMLHtmlElement') should be 'false' and is.
PASS: canGet('targetWindow.HTMLIFrameElement') should be 'false' and is.
PASS: canGet('targetWindow.HTMLImageElement') should be 'false' and is.
PASS: canGet('targetWindow.HTMLInputElement') should be 'false' and is.
PASS: canGet('targetWindow.HTMLIsIndexElement') should be 'false' and is.
PASS: canGet('targetWindow.HTMLLIElement') should be 'false' and is.
PASS: canGet('targetWindow.HTMLLabelElement') should be 'false' and is.
PASS: canGet('targetWindow.HTMLLegendElement') should be 'false' and is.
PASS: canGet('targetWindow.HTMLLinkElement') should be 'false' and is.
PASS: canGet('targetWindow.HTMLMapElement') should be 'false' and is.
PASS: canGet('targetWindow.HTMLMarqueeElement') should be 'false' and is.
PASS: canGet('targetWindow.HTMLMediaElement') should be 'false' and is.
PASS: canGet('targetWindow.HTMLMenuElement') should be 'false' and is.
PASS: canGet('targetWindow.HTMLMetaElement') should be 'false' and is.
PASS: canGet('targetWindow.HTMLModElement') should be 'false' and is.
PASS: canGet('targetWindow.HTMLOListElement') should be 'false' and is.
PASS: canGet('targetWindow.HTMLOptGroupElement') should be 'false' and is.
PASS: canGet('targetWindow.HTMLOptionElement') should be 'false' and is.
PASS: canGet('targetWindow.HTMLParagraphElement') should be 'false' and is.
PASS: canGet('targetWindow.HTMLParamElement') should be 'false' and is.
PASS: canGet('targetWindow.HTMLPreElement') should be 'false' and is.
PASS: canGet('targetWindow.HTMLQuoteElement') should be 'false' and is.
PASS: canGet('targetWindow.HTMLScriptElement') should be 'false' and is.
PASS: canGet('targetWindow.HTMLSelectElement') should be 'false' and is.
PASS: canGet('targetWindow.HTMLSourceElement') should be 'false' and is.
PASS: canGet('targetWindow.HTMLStyleElement') should be 'false' and is.
PASS: canGet('targetWindow.HTMLTableCaptionElement') should be 'false' and is.
PASS: canGet('targetWindow.HTMLTableCellElement') should be 'false' and is.
PASS: canGet('targetWindow.HTMLTableColElement') should be 'false' and is.
PASS: canGet('targetWindow.HTMLTableElement') should be 'false' and is.
PASS: canGet('targetWindow.HTMLTableRowElement') should be 'false' and is.
PASS: canGet('targetWindow.HTMLTableSectionElement') should be 'false' and is.
PASS: canGet('targetWindow.HTMLTextAreaElement') should be 'false' and is.
PASS: canGet('targetWindow.HTMLTitleElement') should be 'false' and is.
PASS: canGet('targetWindow.HTMLUListElement') should be 'false' and is.
PASS: canGet('targetWindow.HTMLVideoElement') should be 'false' and is.
PASS: canGet('targetWindow.Image') should be 'false' and is.
PASS: canGet('targetWindow.MutationEvent') should be 'false' and is.
PASS: canGet('targetWindow.Node') should be 'false' and is.
PASS: canGet('targetWindow.NodeFilter') should be 'false' and is.
PASS: canGet('targetWindow.Option') should be 'false' and is.
PASS: canGet('targetWindow.ProcessingInstruction') should be 'false' and is.
PASS: canGet('targetWindow.Range') should be 'false' and is.
PASS: canGet('targetWindow.RangeError') should be 'false' and is.
PASS: canGet('targetWindow.ReferenceError') should be 'false' and is.
PASS: canGet('targetWindow.SyntaxError') should be 'false' and is.
PASS: canGet('targetWindow.Text') should be 'false' and is.
PASS: canGet('targetWindow.TypeError') should be 'false' and is.
PASS: canGet('targetWindow.URIError') should be 'false' and is.
PASS: canGet('targetWindow.XMLDocument') should be 'false' and is.
PASS: canGet('targetWindow.XMLHttpRequest') should be 'false' and is.
PASS: canGet('targetWindow.XMLSerializer') should be 'false' and is.
PASS: canGet('targetWindow.XPathEvaluator') should be 'false' and is.
PASS: canGet('targetWindow.XPathResult') should be 'false' and is.
PASS: canGet('targetWindow.XSLTProcessor') should be 'false' and is.

----- tests for getting of not allowed Functions -----

PASS: canGet('targetWindow.addEventListener') should be 'false' and is.
PASS: canGet('targetWindow.alert') should be 'false' and is.
PASS: canGet('targetWindow.atob') should be 'false' and is.
PASS: canGet('targetWindow.btoa') should be 'false' and is.
PASS: canGet('targetWindow.captureEvents') should be 'false' and is.
PASS: canGet('targetWindow.clearInterval') should be 'false' and is.
PASS: canGet('targetWindow.clearTimeout') should be 'false' and is.
PASS: canGet('targetWindow.confirm') should be 'false' and is.
PASS: canGet('targetWindow.constructor') should be 'false' and is.
PASS: canGet('targetWindow.find') should be 'false' and is.
PASS: canGet('targetWindow.getComputedStyle') should be 'false' and is.
PASS: canGet('targetWindow.getMatchedCSSRules') should be 'false' and is.
PASS: canGet('targetWindow.getSelection') should be 'false' and is.
PASS: canGet('targetWindow.moveBy') should be 'false' and is.
PASS: canGet('targetWindow.moveTo') should be 'false' and is.
PASS: canGet('targetWindow.open') should be 'false' and is.
PASS: canGet('targetWindow.print') should be 'false' and is.
PASS: canGet('targetWindow.prompt') should be 'false' and is.
PASS: canGet('targetWindow.releaseEvents') should be 'false' and is.
PASS: canGet('targetWindow.removeEventListener') should be 'false' and is.
PASS: canGet('targetWindow.resizeBy') should be 'false' and is.
PASS: canGet('targetWindow.resizeTo') should be 'false' and is.
PASS: canGet('targetWindow.scroll') should be 'false' and is.
PASS: canGet('targetWindow.scrollBy') should be 'false' and is.
PASS: canGet('targetWindow.scrollTo') should be 'false' and is.
PASS: canGet('targetWindow.setInterval') should be 'false' and is.
PASS: canGet('targetWindow.setTimeout') should be 'false' and is.
PASS: canGet('targetWindow.stop') should be 'false' and is.

----- tests for getting of not allowed Attributes -----

PASS: canGet('targetWindow.clientInformation') should be 'false' and is.
PASS: canGet('targetWindow.console') should be 'false' and is.
PASS: canGet('targetWindow.crypto') should be 'false' and is.
PASS: canGet('targetWindow.defaultStatus') should be 'false' and is.
PASS: canGet('targetWindow.defaultstatus') should be 'false' and is.
PASS: canGet('targetWindow.devicePixelRatio') should be 'false' and is.
Firefox allows access to 'document' but throws an exception when you access its properties.
PASS: canGet('targetWindow.document') should be 'false' and is.
PASS: canGet('targetWindow.embeds') should be 'false' and is.
PASS: canGet('targetWindow.eval') should be 'false' and is.
PASS: canGet('targetWindow.event') should be 'false' and is.
PASS: canGet('targetWindow.frameElement') should be 'false' and is.
PASS: canGet('targetWindow.history') should be 'false' and is.
PASS: canGet('targetWindow.images') should be 'false' and is.
PASS: canGet('targetWindow.innerHeight') should be 'false' and is.
PASS: canGet('targetWindow.innerWidth') should be 'false' and is.
PASS: canGet('targetWindow.locationbar') should be 'false' and is.
PASS: canGet('targetWindow.menubar') should be 'false' and is.
PASS: canGet('targetWindow.name') should be 'false' and is.
PASS: canGet('targetWindow.navigator') should be 'false' and is.
PASS: canGet('targetWindow.offscreenBuffering') should be 'false' and is.
PASS: canGet('targetWindow.onabort') should be 'false' and is.
PASS: canGet('targetWindow.onbeforeunload') should be 'false' and is.
PASS: canGet('targetWindow.onblur') should be 'false' and is.
PASS: canGet('targetWindow.onchange') should be 'false' and is.
PASS: canGet('targetWindow.onclick') should be 'false' and is.
PASS: canGet('targetWindow.ondblclick') should be 'false' and is.
PASS: canGet('targetWindow.onerror') should be 'false' and is.
PASS: canGet('targetWindow.onfocus') should be 'false' and is.
PASS: canGet('targetWindow.onkeydown') should be 'false' and is.
PASS: canGet('targetWindow.onkeypress') should be 'false' and is.
PASS: canGet('targetWindow.onkeyup') should be 'false' and is.
PASS: canGet('targetWindow.onload') should be 'false' and is.
PASS: canGet('targetWindow.onmousedown') should be 'false' and is.
PASS: canGet('targetWindow.onmousemove') should be 'false' and is.
PASS: canGet('targetWindow.onmouseout') should be 'false' and is.
PASS: canGet('targetWindow.onmouseover') should be 'false' and is.
PASS: canGet('targetWindow.onmouseup') should be 'false' and is.
PASS: canGet('targetWindow.onmousewheel') should be 'false' and is.
PASS: canGet('targetWindow.onreset') should be 'false' and is.
PASS: canGet('targetWindow.onresize') should be 'false' and is.
PASS: canGet('targetWindow.onscroll') should be 'false' and is.
PASS: canGet('targetWindow.onsearch') should be 'false' and is.
PASS: canGet('targetWindow.onselect') should be 'false' and is.
PASS: canGet('targetWindow.onsubmit') should be 'false' and is.
PASS: canGet('targetWindow.onunload') should be 'false' and is.
PASS: canGet('targetWindow.outerHeight') should be 'false' and is.
PASS: canGet('targetWindow.outerWidth') should be 'false' and is.
PASS: canGet('targetWindow.pageXOffset') should be 'false' and is.
PASS: canGet('targetWindow.pageYOffset') should be 'false' and is.
PASS: canGet('targetWindow.personalbar') should be 'false' and is.
PASS: canGet('targetWindow.plugins') should be 'false' and is.
PASS: canGet('targetWindow.screen') should be 'false' and is.
PASS: canGet('targetWindow.screenLeft') should be 'false' and is.
PASS: canGet('targetWindow.screenTop') should be 'false' and is.
PASS: canGet('targetWindow.screenX') should be 'false' and is.
PASS: canGet('targetWindow.screenY') should be 'false' and is.
PASS: canGet('targetWindow.scrollX') should be 'false' and is.
PASS: canGet('targetWindow.scrollY') should be 'false' and is.
PASS: canGet('targetWindow.scrollbars') should be 'false' and is.
PASS: canGet('targetWindow.status') should be 'false' and is.
PASS: canGet('targetWindow.statusbar') should be 'false' and is.
PASS: canGet('targetWindow.toolbar') should be 'false' and is.
